User-Feedback Function Added
Users now see selection buttons for each uploaded poster for human or ai-made. When the user selects their answer, they are informed if they answered correctly. The button is then disabled from selection again during the session. Adjustments are needed for dark mode. Otherwise this is functional as described in the design brief